continue cleaning up chain package and refactoring e2e
