+++ tests/e2e/chain/config.go
@@ -1,9 +1,23 @@
 package chain
 
 import (
+	"encoding/json"
 	"fmt"
+	"path/filepath"
+	"strings"
 
+	"github.com/cosmos/cosmos-sdk/server"
+	srvconfig "github.com/cosmos/cosmos-sdk/server/config"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/cosmos/cosmos-sdk/x/genutil"
+	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+	"github.com/spf13/viper"
+	tmconfig "github.com/tendermint/tendermint/config"
+	tmjson "github.com/tendermint/tendermint/libs/json"
+
+	"github.com/osmosis-labs/osmosis/v7/tests/e2e/util"
 )
 
 const (
@@ -34,3 +48,255 @@ var (
 	InitBalanceStrA = fmt.Sprintf("%d%s,%d%s", OsmoBalanceA, OsmoDenom, StakeBalanceA, StakeDenom)
 	InitBalanceStrB = fmt.Sprintf("%d%s,%d%s", OsmoBalanceB, OsmoDenom, StakeBalanceB, StakeDenom)
 )
+
+func initValidatorConfigs(c *Chain) error {
+	for i, val := range c.Validators {
+		tmCfgPath := filepath.Join(val.ConfigDir(), "config", "config.toml")
+
+		vpr := viper.New()
+		vpr.SetConfigFile(tmCfgPath)
+		if err := vpr.ReadInConfig(); err != nil {
+			return err
+		}
+
+		valConfig := &tmconfig.Config{}
+		if err := vpr.Unmarshal(valConfig); err != nil {
+			return err
+		}
+
+		valConfig.P2P.ListenAddress = "tcp://0.0.0.0:26656"
+		valConfig.P2P.AddrBookStrict = false
+		valConfig.P2P.ExternalAddress = fmt.Sprintf("%s:%d", val.InstanceName(), 26656)
+		valConfig.RPC.ListenAddress = "tcp://0.0.0.0:26657"
+		valConfig.StateSync.Enable = false
+		valConfig.LogLevel = "info"
+
+		var peers []string
+
+		for j := 0; j < len(c.Validators); j++ {
+			if i == j {
+				continue
+			}
+
+			peer := c.Validators[j]
+			peerID := fmt.Sprintf("%s@%s%d:26656", peer.getNodeKey().ID(), peer.GetMoniker(), j)
+			peers = append(peers, peerID)
+		}
+
+		valConfig.P2P.PersistentPeers = strings.Join(peers, ",")
+
+		tmconfig.WriteConfigFile(tmCfgPath, valConfig)
+
+		// set application configuration
+		appCfgPath := filepath.Join(val.ConfigDir(), "config", "app.toml")
+
+		appConfig := srvconfig.DefaultConfig()
+		appConfig.API.Enable = true
+		appConfig.MinGasPrices = fmt.Sprintf("%s%s", MinGasPrice, OsmoDenom)
+
+		srvconfig.WriteConfigFile(appCfgPath, appConfig)
+	}
+	return nil
+}
+
+func initGenesis(c *Chain) error {
+	_, cdc := util.InitEncodingConfigAndCdc()
+
+	serverCtx := server.NewDefaultContext()
+	config := serverCtx.Config
+
+	config.SetRoot(c.Validators[0].ConfigDir())
+	config.Moniker = c.Validators[0].GetMoniker()
+
+	genFilePath := config.GenesisFile()
+	appGenState, genDoc, err := genutiltypes.GenesisStateFromGenFile(genFilePath)
+	if err != nil {
+		return err
+	}
+
+	var bankGenState banktypes.GenesisState
+	if err := cdc.UnmarshalJSON(appGenState[banktypes.ModuleName], &bankGenState); err != nil {
+		return err
+	}
+
+	bankGenState.DenomMetadata = append(bankGenState.DenomMetadata, banktypes.Metadata{
+		Description: "An example stable token",
+		Display:     OsmoDenom,
+		Base:        OsmoDenom,
+		Symbol:      OsmoDenom,
+		Name:        OsmoDenom,
+		DenomUnits: []*banktypes.DenomUnit{
+			{
+				Denom:    OsmoDenom,
+				Exponent: 0,
+			},
+		},
+	})
+
+	bz, err := cdc.MarshalJSON(&bankGenState)
+	if err != nil {
+		return err
+	}
+	appGenState[banktypes.ModuleName] = bz
+
+	var genUtilGenState genutiltypes.GenesisState
+	if err := cdc.UnmarshalJSON(appGenState[genutiltypes.ModuleName], &genUtilGenState); err != nil {
+		return err
+	}
+
+	// generate genesis txs
+	genTxs := make([]json.RawMessage, len(c.Validators))
+	for i, val := range c.Validators {
+		stakeAmountCoin := StakeAmountCoinA
+		if c.Id != ChainAID {
+			stakeAmountCoin = StakeAmountCoinB
+		}
+		createValmsg, err := val.buildCreateValidatorMsg(stakeAmountCoin)
+		if err != nil {
+			return err
+		}
+
+		signedTx, err := val.signMsg(createValmsg)
+		if err != nil {
+			return err
+		}
+
+		txRaw, err := cdc.MarshalJSON(signedTx)
+		if err != nil {
+			return err
+		}
+
+		genTxs[i] = txRaw
+	}
+
+	genUtilGenState.GenTxs = genTxs
+
+	bz, err = cdc.MarshalJSON(&genUtilGenState)
+	if err != nil {
+		return err
+	}
+	appGenState[genutiltypes.ModuleName] = bz
+
+	bz, err = json.MarshalIndent(appGenState, "", "  ")
+	if err != nil {
+		return err
+	}
+
+	genDoc.AppState = bz
+
+	bz, err = tmjson.MarshalIndent(genDoc, "", "  ")
+	if err != nil {
+		return err
+	}
+
+	// write the updated genesis file to each validator
+	for _, val := range c.Validators {
+		if err := util.WriteFile(filepath.Join(val.ConfigDir(), "config", "genesis.json"), bz); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func initNodes(c *Chain) error {
+	if err := c.createAndInitValidators(2); err != nil {
+		return err
+	}
+
+	// initialize a genesis file for the first validator
+	val0ConfigDir := c.Validators[0].ConfigDir()
+	for _, val := range c.Validators {
+		if c.Id == ChainAID {
+			if err := addAccount(val0ConfigDir, "", InitBalanceStrA, val.GetKeyInfo().GetAddress()); err != nil {
+				return err
+			}
+		} else if c.Id == ChainBID {
+			if err := addAccount(val0ConfigDir, "", InitBalanceStrB, val.GetKeyInfo().GetAddress()); err != nil {
+				return err
+			}
+		}
+	}
+
+	// copy the genesis file to the remaining validators
+	for _, val := range c.Validators[1:] {
+		_, err := util.CopyFile(
+			filepath.Join(val0ConfigDir, "config", "genesis.json"),
+			filepath.Join(val.ConfigDir(), "config", "genesis.json"),
+		)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func addAccount(path, moniker, amountStr string, accAddr sdk.AccAddress) error {
+	serverCtx := server.NewDefaultContext()
+	config := serverCtx.Config
+
+	config.SetRoot(path)
+	config.Moniker = moniker
+
+	coins, err := sdk.ParseCoinsNormalized(amountStr)
+	if err != nil {
+		return fmt.Errorf("failed to parse coins: %w", err)
+	}
+
+	balances := banktypes.Balance{Address: accAddr.String(), Coins: coins.Sort()}
+	genAccount := authtypes.NewBaseAccount(accAddr, nil, 0, 0)
+
+	genFile := config.GenesisFile()
+	appState, genDoc, err := genutiltypes.GenesisStateFromGenFile(genFile)
+	if err != nil {
+		return fmt.Errorf("failed to unmarshal genesis state: %w", err)
+	}
+
+	authGenState := authtypes.GetGenesisStateFromAppState(util.Cdc, appState)
+
+	accs, err := authtypes.UnpackAccounts(authGenState.Accounts)
+	if err != nil {
+		return fmt.Errorf("failed to get accounts from any: %w", err)
+	}
+
+	if accs.Contains(accAddr) {
+		return fmt.Errorf("failed to add account to genesis state; account already exists: %s", accAddr)
+	}
+
+	// Add the new account to the set of genesis accounts and sanitize the
+	// accounts afterwards.
+	accs = append(accs, genAccount)
+	accs = authtypes.SanitizeGenesisAccounts(accs)
+
+	genAccs, err := authtypes.PackAccounts(accs)
+	if err != nil {
+		return fmt.Errorf("failed to convert accounts into any's: %w", err)
+	}
+
+	authGenState.Accounts = genAccs
+
+	authGenStateBz, err := util.Cdc.MarshalJSON(&authGenState)
+	if err != nil {
+		return fmt.Errorf("failed to marshal auth genesis state: %w", err)
+	}
+
+	appState[authtypes.ModuleName] = authGenStateBz
+
+	bankGenState := banktypes.GetGenesisStateFromAppState(util.Cdc, appState)
+	bankGenState.Balances = append(bankGenState.Balances, balances)
+	bankGenState.Balances = banktypes.SanitizeGenesisBalances(bankGenState.Balances)
+
+	bankGenStateBz, err := util.Cdc.MarshalJSON(bankGenState)
+	if err != nil {
+		return fmt.Errorf("failed to marshal bank genesis state: %w", err)
+	}
+
+	appState[banktypes.ModuleName] = bankGenStateBz
+
+	appStateJSON, err := json.Marshal(appState)
+	if err != nil {
+		return fmt.Errorf("failed to marshal application genesis state: %w", err)
+	}
+
+	genDoc.AppState = appStateJSON
+	return genutil.ExportGenesisFile(genDoc, genFile)
+}
